externalized many algo customizations for generated python + added hook for property examples
