--- h2o-bindings/bin/custom/gen_gbm.py
@@ -1,29 +1,0 @@
-header = """
-my header
-"""
-
-doc_class = """
-more doc
-"""
-
-doc_property1 = """
-doc for property1
-"""
-
-
-def example_class():
-    return "this should be executable code showing how to use the gbm class"
-
-
-def example_property1():
-    return "this should also be executable code showing how to use property1"
-
-
-def extras():
-
-    def new_method(self, prop=None):
-        pass
-
-
-
-
