+++ h2o-bindings/bin/custom/python/gen_drf.py
@@ -1,0 +1,1 @@
+deprecated_attributes = ['offset_column', 'distribution']
