+++ h2o-bindings/bin/custom/python/gen_glm.py
@@ -1,0 +1,65 @@
+imports = """import h2o"""
+
+class_doc = """
+Fits a generalized linear model, specified by a response variable, a set of predictors, and a
+description of the error distribution.
+
+A subclass of :class:`ModelBase` is returned. The specific subclass depends on the machine learning task
+at hand (if it's binomial classification, then an H2OBinomialModel is returned, if it's regression then a
+H2ORegressionModel is returned). The default print-out of the models is shown, but further GLM-specific
+information can be queried out of the object. Upon completion of the GLM, the resulting object has
+coefficients, normalized coefficients, residual/null deviance, aic, and a host of model metrics including
+MSE, AUC (for logistic regression), degrees of freedom, and confusion matrices.
+"""
+
+class_extras = """
+@property
+def Lambda(self):
+    \"""DEPRECATED. Use ``self.lambda_`` instead\"""
+    return self._parms["lambda"] if "lambda" in self._parms else None
+
+@Lambda.setter
+def Lambda(self, value):
+    self._parms["lambda"] = value
+
+@staticmethod
+def getGLMRegularizationPath(model):
+    \"""
+    Extract full regularization path explored during lambda search from glm model.
+
+    :param model: source lambda search model
+    \"""
+    x = h2o.api("GET /3/GetGLMRegPath", data={"model": model._model_json["model_id"]["name"]})
+    ns = x.pop("coefficient_names")
+    res = {
+        "lambdas": x["lambdas"],
+        "explained_deviance_train": x["explained_deviance_train"],
+        "explained_deviance_valid": x["explained_deviance_valid"],
+        "coefficients": [dict(zip(ns, y)) for y in x["coefficients"]],
+    }
+    if "coefficients_std" in x:
+        res["coefficients_std"] = [dict(zip(ns, y)) for y in x["coefficients_std"]]
+    return res
+
+@staticmethod
+def makeGLMModel(model, coefs, threshold=.5):
+    \"""
+    Create a custom GLM model using the given coefficients.
+
+    Needs to be passed source model trained on the dataset to extract the dataset information from.
+
+    :param model: source model, used for extracting dataset information
+    :param coefs: dictionary containing model coefficients
+    :param threshold: (optional, only for binomial) decision threshold used for classification
+    \"""
+    model_json = h2o.api(
+        "POST /3/MakeGLMModel",
+        data={"model": model._model_json["model_id"]["name"],
+              "names": list(coefs.keys()),
+              "beta": list(coefs.values()),
+              "threshold": threshold}
+    )
+    m = H2OGeneralizedLinearEstimator()
+    m._resolve_model(model_json["model_id"]["name"], model_json)
+    return m
+"""
