+++ h2o-bindings/bin/custom/python/gen_glrm.py
@@ -1,0 +1,6 @@
+class_doc = """Builds a generalized low rank model of a H2O dataset."""
+
+class_init_extra = """
+self._parms["_rest_version"] = 3
+"""
+
