+++ h2o-bindings/bin/custom/python/gen_kmeans.py
@@ -1,0 +1,2 @@
+class_doc = """Performs k-means clustering on an H2O dataset."""
+
