+++ h2o-bindings/bin/custom/python/gen_svd.py
@@ -1,0 +1,5 @@
+
+class_init_extra = """
+self._parms["_rest_version"] = 99
+"""
+
