+++ h2o-bindings/bin/gen_python.py
@@ -2,7 +2,7 @@
 # -*- encoding: utf-8 -*-
 from __future__ import unicode_literals
 import bindings as bi
-import sys
+import inspect, os, sys
 PY3 = sys.version_info[0] == 3
 str_type = str if PY3 else (str, unicode)
 
@@ -119,12 +119,12 @@ reserved_words = {
 # ----------------------------------------------------------------------------------------------------------------------
 def gen_module(schema, algo):
     classname = algo_to_classname(algo)
-    extra_imports = extra_imports_for(algo)
-    help_preamble = help_preamble_for(algo)
-    help_epilogue = help_epilogue_for(algo)
-    init_extra = init_extra_for(algo)
-    class_extra = class_extra_for(algo)
-    module_extra = module_extra_for(algo)
+    extra_imports = get_customizations_for(algo, 'imports')
+    class_doc = get_customizations_for(algo, 'class_doc')
+    class_examples = get_customizations_for(algo, 'class_examples')
+    class_init_extra = get_customizations_for(algo, 'class_init_extra')
+    class_extras = get_customizations_for(algo, 'class_extras')
+    module_extras = get_customizations_for(algo, 'module_extras')
 
     param_names = []
     for param in schema["parameters"]:
@@ -163,27 +163,23 @@ def gen_module(schema, algo):
     yield "from h2o.estimators.estimator_base import H2OEstimator"
     yield "from h2o.exceptions import H2OValueError"
     yield "from h2o.frame import H2OFrame"
-    if classname == "H2OStackedEnsembleEstimator":
-        yield "from h2o.utils.shared_utils import quoted"
-        yield "from h2o.utils.typechecks import assert_is_type, Enum, numeric, is_type"
-        yield "import json"
-        yield "import ast"
-    else:
-        yield "from h2o.utils.typechecks import assert_is_type, Enum, numeric"
+    yield "from h2o.utils.typechecks import assert_is_type, Enum, numeric"
     if extra_imports:
-        yield reindent_block(extra_imports, 0) + ""
+        yield reindent_block(extra_imports, 0)
     yield ""
     yield ""
     yield "class %s(H2OEstimator):" % classname
-    yield "    \"\"\""
+    yield '    """'
     yield "    " + schema["algo_full_name"]
     yield ""
-    if help_preamble:
-        yield "    %s" % reindent_block(help_preamble, 4)
-    if help_epilogue:
+    if class_doc:
+        yield "    %s" % reindent_block(class_doc, 4)
+    if class_examples:
         yield ""
-        yield "    %s" % reindent_block(help_epilogue, 4)
-    yield "    \"\"\""
+        yield "    :examples:"
+        yield ""
+        yield "    %s" % reindent_block(class_examples, 4)
+    yield '    """'
     yield ""
     yield '    algo = "%s"' % algo
     yield ""
@@ -210,10 +206,8 @@ def gen_module(schema, algo):
     yield "                setattr(self, pname, pvalue)"
     yield "            else:"
     yield '                raise H2OValueError("Unknown parameter %s = %r" % (pname, pvalue))'
-    if algo == "svd":
-        yield "        self._parms['_rest_version'] = 99"
-    if init_extra:
-        yield "        " + reindent_block(init_extra, 8)
+    if class_init_extra:
+        yield "        " + reindent_block(class_init_extra, 8)
     yield ""
     for param in schema["parameters"]:
         pname = param["pname"]
@@ -223,30 +217,39 @@ def gen_module(schema, algo):
 
         if param["dtype"].startswith("Enum"):
             vals = param["dtype"][5:-1].split(", ")
-            extrahelp = "One of: " + ", ".join("``%s``" % v for v in vals)
+            property_doc = "One of: " + ", ".join("``%s``" % v for v in vals)
         else:
             if pname == "metalearner_params":
-                extrahelp = "Type: ``dict``"
+                property_doc = "Type: ``dict``"
             else:
-                extrahelp = "Type: ``%s``" % param["dtype"]
+                property_doc = "Type: ``%s``" % param["dtype"]
         if param["default_value"] is None:
-            extrahelp += "."
+            property_doc += "."
         else:
             if pname == "metalearner_params":
-                extrahelp += "  (default: ``None``)."
+                property_doc += "  (default: ``None``)."
             else:
-                extrahelp += "  (default: ``%s``)." % stringify(param["default_value"])
+                property_doc += "  (default: ``%s``)." % stringify(param["default_value"])
 
-        deprecated = (algo == 'drf' and pname in ['offset_column', 'distribution'])
+        deprecated = pname in get_customizations_for(algo, 'deprecated_attributes', [])
         yield "    @property"
         yield "    def %s(self):" % pname
         yield '        """'
         yield "        %s%s" % ("[Deprecated] " if deprecated else "", bi.wrap(param["help"], indent=(" " * 8), indent_first=False))
         yield ""
-        yield "        %s" % bi.wrap(extrahelp, indent=(" " * 8), indent_first=False)
-        if pname == "metalearner_params":
-            yield "        Example: metalearner_gbm_params = {'max_depth': 2, 'col_sample_rate': 0.3}"
+        yield "        %s" % bi.wrap(property_doc, indent=(" " * 8), indent_first=False)
+        property_doc = get_customizations_for(algo, "property_"+pname+"_doc")
+        if property_doc:
+            yield ""
+            yield "        %s" % reindent_block(property_doc, 8)
+        property_examples = get_customizations_for(algo, "property_"+pname+"_examples")
+        if property_examples:
+            yield ""
+            yield "        :examples:"
+            yield ""
+            yield "        %s" % reindent_block(property_examples, 8)
         yield '        """'
+
         if pname != "metalearner_params":
             yield "        return self._parms.get(\"%s\")" % sname
         else:
@@ -292,12 +295,11 @@ def gen_module(schema, algo):
             yield "        self._parms[\"%s\"] = %s" % (sname, pname)
         yield ""
         yield ""
-    if class_extra:
+    if class_extras:
+        yield "    " + reindent_block(class_extras, 4)
+    if module_extras:
         yield ""
-        yield "    " + reindent_block(class_extra, 4)
-    if module_extra:
-        yield ""
-        yield reindent_block(module_extra, 0)
+        yield reindent_block(module_extras, 0)
 
 
 def algo_to_classname(algo):
@@ -317,369 +319,6 @@ def algo_to_classname(algo):
     if algo == "isolationforest": return "H2OIsolationForestEstimator"
     if algo == "psvm": return "H2OSupportVectorMachineEstimator"
     return "H2O" + algo.capitalize() + "Estimator"
-
-
-def extra_imports_for(algo):
-    if algo == "glm" or algo == "deepwater" or algo == "xgboost":
-        return "import h2o"
-
-
-def help_preamble_for(algo):
-    if algo == "coxph":
-        return """
-            Trains a Cox Proportional Hazards Model (CoxPH) on an H2O dataset"""
-    if algo == "deeplearning":
-        return """
-            Build a Deep Neural Network model using CPUs
-            Builds a feed-forward multilayer artificial neural network on an H2OFrame"""
-    if algo == "deepwater":
-        return """
-            Build a Deep Learning model using multiple native GPU backends
-            Builds a deep neural network on an H2OFrame containing various data sources"""
-    if algo == "kmeans":
-        return """Performs k-means clustering on an H2O dataset."""
-    if algo == "glrm":
-        return """Builds a generalized low rank model of a H2O dataset."""
-    if algo == "glm":
-        return """
-            Fits a generalized linear model, specified by a response variable, a set of predictors, and a
-            description of the error distribution."""
-    if algo == "gbm":
-        return """
-            Builds gradient boosted trees on a parsed data set, for regression or classification.
-            The default distribution function will guess the model type based on the response column type.
-            Otherwise, the response column must be an enum for "bernoulli" or "multinomial", and numeric
-            for all other distributions."""
-    if algo == "xgboost":
-        return """Builds a eXtreme Gradient Boosting model using the native XGBoost backend."""
-    if algo == "naivebayes":
-        return """
-            The naive Bayes classifier assumes independence between predictor variables
-            conditional on the response, and a Gaussian distribution of numeric predictors with
-            mean and standard deviation computed from the training dataset. When building a naive
-            Bayes classifier, every row in the training dataset that contains at least one NA will
-            be skipped completely. If the test dataset has missing values, then those predictors
-            are omitted in the probability calculation during prediction."""
-    if algo == "stackedensemble":
-        return """
-            Builds a stacked ensemble (aka "super learner") machine learning method that uses two
-            or more H2O learning algorithms to improve predictive performance. It is a loss-based
-            supervised learning method that finds the optimal combination of a collection of prediction
-            algorithms.This method supports regression and binary classification. """
-    if algo == "isolationforest":
-        return """
-            Builds an Isolation Forest model. Isolation Forest algorithm samples the training frame
-            and in each iteration builds a tree that partitions the space of the sample observations until
-            it isolates each observation. Length of the path from root to a leaf node of the resulting tree
-            is used to calculate the anomaly score. Anomalies are easier to isolate and their average
-            tree path is expected to be shorter than paths of regular observations.
-        """
-
-
-def help_epilogue_for(algo):
-    if algo == "deeplearning":
-        return """Examples
-                       --------
-                         >>> import h2o
-                         >>> from h2o.estimators.deeplearning import H2ODeepLearningEstimator
-                         >>> h2o.connect()
-                         >>> rows = [[1,2,3,4,0], [2,1,2,4,1], [2,1,4,2,1], [0,1,2,34,1], [2,3,4,1,0]] * 50
-                         >>> fr = h2o.H2OFrame(rows)
-                         >>> fr[4] = fr[4].asfactor()
-                         >>> model = H2ODeepLearningEstimator()
-                         >>> model.train(x=range(4), y=4, training_frame=fr)"""
-    if algo == "stackedensemble":
-        return """Examples
-                       --------
-                         >>> import h2o
-                         >>> h2o.init()
-                         >>> from h2o.estimators.random_forest import H2ORandomForestEstimator
-                         >>> from h2o.estimators.gbm import H2OGradientBoostingEstimator
-                         >>> from h2o.estimators.stackedensemble import H2OStackedEnsembleEstimator
-                         >>> col_types = ["numeric", "numeric", "numeric", "enum", "enum", "numeric", "numeric", "numeric", "numeric"]
-                         >>> data = h2o.import_file("http://h2o-public-test-data.s3.amazonaws.com/smalldata/prostate/prostate.csv", col_types=col_types)
-                         >>> train, test = data.split_frame(ratios=[.8], seed=1)
-                         >>> x = ["CAPSULE","GLEASON","RACE","DPROS","DCAPS","PSA","VOL"]
-                         >>> y = "AGE"
-                         >>> nfolds = 5
-                         >>> my_gbm = H2OGradientBoostingEstimator(nfolds=nfolds, fold_assignment="Modulo", keep_cross_validation_predictions=True)
-                         >>> my_gbm.train(x=x, y=y, training_frame=train)
-                         >>> my_rf = H2ORandomForestEstimator(nfolds=nfolds, fold_assignment="Modulo", keep_cross_validation_predictions=True)
-                         >>> my_rf.train(x=x, y=y, training_frame=train)
-                         >>> stack = H2OStackedEnsembleEstimator(model_id="my_ensemble", training_frame=train, validation_frame=test, base_models=[my_gbm.model_id, my_rf.model_id])
-                         >>> stack.train(x=x, y=y, training_frame=train, validation_frame=test)
-                         >>> stack.model_performance()"""
-    if algo == "glm":
-        return """
-            A subclass of :class:`ModelBase` is returned. The specific subclass depends on the machine learning task
-            at hand (if it's binomial classification, then an H2OBinomialModel is returned, if it's regression then a
-            H2ORegressionModel is returned). The default print-out of the models is shown, but further GLM-specific
-            information can be queried out of the object. Upon completion of the GLM, the resulting object has
-            coefficients, normalized coefficients, residual/null deviance, aic, and a host of model metrics including
-            MSE, AUC (for logistic regression), degrees of freedom, and confusion matrices."""
-
-
-def init_extra_for(algo):
-    if algo == "deeplearning":
-        return "if isinstance(self, H2OAutoEncoderEstimator): self._parms['autoencoder'] = True"
-    if algo == "glrm":
-        return """self._parms["_rest_version"] = 3"""
-    if algo == "stackedensemble":
-        return """self._parms["_rest_version"] = 99"""
-    if algo == "aggregator":
-        return """self._parms["_rest_version"] = 99"""
-
-
-def class_extra_for(algo):
-    if algo == "aggregator":
-        # Add aggregated_frame to model
-        return """
-            @property
-            def aggregated_frame(self):
-                if (self._model_json is not None and
-                    self._model_json.get("output", {}).get("output_frame", {}).get("name") is not None):
-                    out_frame_name = self._model_json["output"]["output_frame"]["name"]
-                    return H2OFrame.get_frame(out_frame_name)"""
-    if algo == "glm":
-        # Before we were replacing .lambda property with .Lambda. However that violates Python naming conventions for
-        # variables, so now we prefer to map that property to .lambda_. The old name remains, for compatibility reasons.
-        return """
-            @property
-            def Lambda(self):
-                \"""DEPRECATED. Use ``self.lambda_`` instead\"""
-                return self._parms["lambda"] if "lambda" in self._parms else None
-
-            @Lambda.setter
-            def Lambda(self, value):
-                self._parms["lambda"] = value
-
-            @staticmethod
-            def getGLMRegularizationPath(model):
-                \"\"\"
-                Extract full regularization path explored during lambda search from glm model.
-
-                :param model: source lambda search model
-                \"\"\"
-                x = h2o.api("GET /3/GetGLMRegPath", data={"model": model._model_json["model_id"]["name"]})
-                ns = x.pop("coefficient_names")
-                res = {
-                    "lambdas": x["lambdas"],
-                    "explained_deviance_train": x["explained_deviance_train"],
-                    "explained_deviance_valid": x["explained_deviance_valid"],
-                    "coefficients": [dict(zip(ns, y)) for y in x["coefficients"]],
-                }
-                if "coefficients_std" in x:
-                    res["coefficients_std"] = [dict(zip(ns, y)) for y in x["coefficients_std"]]
-                return res
-
-            @staticmethod
-            def makeGLMModel(model, coefs, threshold=.5):
-                \"\"\"
-                Create a custom GLM model using the given coefficients.
-
-                Needs to be passed source model trained on the dataset to extract the dataset information from.
-
-                :param model: source model, used for extracting dataset information
-                :param coefs: dictionary containing model coefficients
-                :param threshold: (optional, only for binomial) decision threshold used for classification
-                \"\"\"
-                model_json = h2o.api(
-                    "POST /3/MakeGLMModel",
-                    data={"model": model._model_json["model_id"]["name"],
-                          "names": list(coefs.keys()),
-                          "beta": list(coefs.values()),
-                          "threshold": threshold}
-                )
-                m = H2OGeneralizedLinearEstimator()
-                m._resolve_model(model_json["model_id"]["name"], model_json)
-                return m"""
-
-    elif algo == "deepwater":
-        return """
-        # Ask the H2O server whether a Deep Water model can be built (depends on availability of native backends)
-        @staticmethod
-        def available():
-            \"\"\"Returns True if a deep water model can be built, or False otherwise.\"\"\"
-            builder_json = h2o.api("GET /3/ModelBuilders", data={"algo": "deepwater"})
-            visibility = builder_json["model_builders"]["deepwater"]["visibility"]
-            if visibility == "Experimental":
-                print("Cannot build a Deep Water model - no backend found.")
-                return False
-            else:
-                return True
-        """
-
-    elif algo == "xgboost":
-        return """
-        # Ask the H2O server whether a XGBoost model can be built (depends on availability of native backends)
-        @staticmethod
-        def available():
-            \"\"\"
-            Returns True if a XGBoost model can be built, or False otherwise.
-            \"\"\"
-            if "XGBoost" not in h2o.cluster().list_core_extensions():
-                print("Cannot build an XGBoost model - no backend found.")
-                return False
-            else:
-                return True
-        """
-    elif algo == "stackedensemble":
-        return """
-        # Print the metalearner of an H2OStackedEnsembleEstimator.
-        def metalearner(self):
-            model = self._model_json["output"]
-            if "metalearner" in model and model["metalearner"] is not None:
-                return model["metalearner"]
-            print("No metalearner for this model")  
-              
-        #Fetch the levelone_frame_id for an H2OStackedEnsembleEstimator.   
-        def levelone_frame_id(self):
-            model = self._model_json["output"]
-            if "levelone_frame_id" in model and model["levelone_frame_id"] is not None:
-                return model["levelone_frame_id"]
-            print("No levelone_frame_id for this model")         
-            
-        def stacking_strategy(self):
-            model = self._model_json["output"]
-            if "stacking_strategy" in model and model["stacking_strategy"] is not None:
-                return model["stacking_strategy"]
-            print("No stacking strategy for this model")  
-        
-        # Override train method to support blending 
-        def train(self, x=None, y=None, training_frame=None, blending_frame=None, **kwargs):
-            blending_frame = H2OFrame._validate(blending_frame, 'blending_frame', required=False)
-            
-            def extend_parms(parms):
-                if blending_frame is not None:
-                    parms['blending_frame'] = blending_frame
-                if self.metalearner_fold_column is not None:
-                    parms['ignored_columns'].remove(quoted(self.metalearner_fold_column))
-                    
-            super(self.__class__, self)._train(x, y, training_frame, 
-                                               extend_parms_fn=extend_parms, 
-                                               **kwargs)
-        """
-    elif algo == "word2vec":
-        return """
-        def _requires_training_frame(self):
-            \"\"\"
-            Determines if Word2Vec algorithm requires a training frame.
-            :return: False.
-            \"\"\"
-            return False
-
-        @staticmethod
-        def from_external(external=H2OFrame):
-            \"\"\"
-            Creates new H2OWord2vecEstimator based on an external model.
-            :param external: H2OFrame with an external model
-            :return: H2OWord2vecEstimator instance representing the external model
-            \"\"\"
-            w2v_model = H2OWord2vecEstimator(pre_trained=external)
-            w2v_model.train()
-            return w2v_model
-
-        def _determine_vec_size(self):
-            \"\"\"
-            Determines vec_size for a pre-trained model after basic model verification.
-            \"\"\"
-            first_column = self.pre_trained.types[self.pre_trained.columns[0]]
-
-            if first_column != 'string':
-                raise H2OValueError("First column of given pre_trained model %s is required to be a String",
-                                    self.pre_trained.frame_id)
-
-            if list(self.pre_trained.types.values()).count('string') > 1:
-                raise H2OValueError("There are multiple columns in given pre_trained model %s with a String type.",
-                                    self.pre_trained.frame_id)
-
-            self.vec_size = self.pre_trained.dim[1] - 1;
-        """
-    elif algo == "pca":
-        return """
-        def init_for_pipeline(self):
-            \"\"\"
-            Returns H2OPCA object which implements fit and transform method to be used in sklearn.Pipeline properly.
-            All parameters defined in self.__params, should be input parameters in H2OPCA.__init__ method.
-
-            :returns: H2OPCA object
-            \"\"\"
-            import inspect
-            from h2o.transforms.decomposition import H2OPCA
-            # check which parameters can be passed to H2OPCA init
-            var_names = list(dict(inspect.getmembers(H2OPCA.__init__.__code__))['co_varnames'])
-            parameters = {k: v for k, v in self._parms.items() if k in var_names}
-            return H2OPCA(**parameters)
-        """
-    if algo == "svd":
-        return """
-        def init_for_pipeline(self):
-            \"\"\"
-            Returns H2OSVD object which implements fit and transform method to be used in sklearn.Pipeline properly.
-            All parameters defined in self.__params, should be input parameters in H2OSVD.__init__ method.
-
-            :returns: H2OSVD object
-            \"\"\"
-            import inspect
-            from h2o.transforms.decomposition import H2OSVD
-            # check which parameters can be passed to H2OSVD init
-            var_names = list(dict(inspect.getmembers(H2OSVD.__init__.__code__))['co_varnames'])
-            parameters = {k: v for k, v in self._parms.items() if k in var_names}
-            return H2OSVD(**parameters)
-        """
-    elif algo == "coxph":
-        return """
-        def _additional_used_columns(self, parms):
-            \"\"\"
-            :return: Start and stop column if specified.
-            \"\"\"
-            result = []
-            for col in ["start_column", "stop_column"]:
-                if col in parms and parms[col] is not None:
-                    result.append(parms[col])
-            return result
-        """
-    elif algo == "generic":
-        return """
-        def _requires_training_frame(self):
-            \"\"\"
-            Determines if Generic model requires a training frame.
-            :return: False.
-            \"\"\"
-            return False
-        
-        @staticmethod
-        def from_file(file=str):
-            \"\"\"
-            Creates new Generic model by loading existing embedded model into library, e.g. from H2O MOJO.
-            The imported model must be supported by H2O.
-            :param file: A string containing path to the file to create the model from
-            :return: H2OGenericEstimator instance representing the generic model
-            \"\"\"
-            model = H2OGenericEstimator(path = file)
-            model.train()
-            
-            return model
-        """
-
-
-def module_extra_for(algo):
-    if algo == "deeplearning":
-        return """
-            class H2OAutoEncoderEstimator(H2ODeepLearningEstimator):
-                \"\"\"
-                Examples
-                --------
-                  >>> import h2o as ml
-                  >>> from h2o.estimators.deeplearning import H2OAutoEncoderEstimator
-                  >>> ml.init()
-                  >>> rows = [[1,2,3,4,0]*50, [2,1,2,4,1]*50, [2,1,4,2,1]*50, [0,1,2,34,1]*50, [2,3,4,1,0]*50]
-                  >>> fr = ml.H2OFrame(rows)
-                  >>> fr[4] = fr[4].asfactor()
-                  >>> model = H2OAutoEncoderEstimator()
-                  >>> model.train(x=range(4), training_frame=fr)
-                \"\"\"
-                pass"""
 
 
 def gen_init(modules):
@@ -727,6 +366,24 @@ def gen_models_docs(modules):
             yield "    :members:"
             yield ""
 
+
+_gen_customizations = dict()
+
+
+def get_customizations_for(algo, property=None, default=None):
+    if algo not in _gen_customizations:
+        custom_file = os.path.join(os.path.dirname(__file__), 'custom', 'python', 'gen_'+algo.lower()+'.py')
+        customizations = dict()
+        if os.path.isfile(custom_file):
+            with open(custom_file) as f:
+                exec(f.read(), customizations)
+        _gen_customizations.update({algo: customizations})
+
+    customizations = _gen_customizations[algo]
+    if property is None:
+        return customizations
+    else:
+        return customizations.get(property, default)
 
 # ----------------------------------------------------------------------------------------------------------------------
 #   MAIN:
