+++ h2o-py/h2o/estimators/xgboost.py
@@ -960,12 +960,11 @@ class H2OXGBoostEstimator(H2OEstimator):
         self._parms["gpu_id"] = gpu_id
 
 
-
-    # Ask the H2O server whether a XGBoost model can be built (depends on availability of native backends)
     @staticmethod
     def available():
         """
-        Returns True if a XGBoost model can be built, or False otherwise.
+        Ask the H2O server whether a XGBoost model can be built (depends on availability of native backends).
+        :return: True if a XGBoost model can be built, or False otherwise.
         """
         if "XGBoost" not in h2o.cluster().list_core_extensions():
             print("Cannot build an XGBoost model - no backend found.")
